requirements.txt
mlx>=0.0.10
mlx_clip @ git+https://github.com/harperreed/mlx_clip.git
ocrmac>=1.0.0
qdrant-client>=1.7.0
deepface>=0.0.83
opencv-python-headless
pillow
psutil
numpy

